+++ src/core_modules/capture-core/components/ScopeSelector/hooks/useResetProgramId.js
@@ -1,0 +1,16 @@
+// @flow
+import { useHistory } from 'react-router';
+import { useSelector } from 'react-redux';
+import { urlArguments, getUrlQueries } from '../../../utils/url';
+
+export const useResetProgramId = () => {
+    const history = useHistory();
+    const pathname: string = useSelector(({ router: { location } }) => location.pathname);
+
+    const resetProgramId = (pageToPush: string = pathname) => {
+        const { programId, ...restOfQueries } = getUrlQueries();
+        history.push(`${pageToPush}?${urlArguments({ ...restOfQueries })}`);
+    };
+
+    return { resetProgramId };
+};