+++ src/core_modules/capture-core/components/WorkingLists/TeiWorkingLists/teiWorkingLists.types.js
@@ -1,0 +1,7 @@
+
+// @flow
+export type Props = {|
+    programId: string,
+    orgUnitId: string,
+    ...CssClasses,
+|};