+++ src/core_modules/capture-core/rules/actionsCreator/rulesEngineActionsCreator.js
@@ -2,11 +2,6 @@
 /**
  * @module rulesEngineActionsCreator
  */
-import type { Program, RenderFoundation, TrackerProgram, ProgramStage } from '../../metaData';
-import { runRulesForSingleEvent } from './runRulesForSingleEvent';
-import { runRulesForTEI } from './runRulesForTEI';
-import { postProcessRulesEffects } from './postProcessRulesEffects';
-import { updateRulesEffects } from './rulesEngine.actions';
 import type {
     OutputEffects,
     Enrollment,
@@ -14,8 +9,12 @@ import type {
     OrgUnit,
     EventData,
     EventsData,
-} from '../engine';
-
+} from 'capture-core-utils/rulesEngine';
+import type { Program, RenderFoundation, TrackerProgram, ProgramStage } from '../../metaData';
+import { runRulesForSingleEvent } from './runRulesForSingleEvent';
+import { runRulesForTEI } from './runRulesForTEI';
+import { postProcessRulesEffects } from './postProcessRulesEffects';
+import { updateRulesEffects } from './rulesEngine.actions';
 
 function getRulesActions(
     rulesEffects: ?OutputEffects,