+++ CHANGELOG.md
@@ -1,3 +1,72 @@
+# [1.39.0](https://github.com/dhis2/capture-app/compare/v1.38.0...v1.39.0) (2021-08-16)
+
+
+### Features
+
+* [TECH-633] Update event in redux after edit ([#1917](https://github.com/dhis2/capture-app/issues/1917)) ([cdb8a5c](https://github.com/dhis2/capture-app/commit/cdb8a5ceeeff00afa83323845a283855b31fe61e))
+
+# [1.38.0](https://github.com/dhis2/capture-app/compare/v1.37.2...v1.38.0) (2021-08-13)
+
+
+### Bug Fixes
+
+* show error and warning widgets ([56b1539](https://github.com/dhis2/capture-app/commit/56b1539d677f4dc1e4f4b8137cc2e0c89df807d7))
+* wrong parameter order ([e8fe5ef](https://github.com/dhis2/capture-app/commit/e8fe5ef6d262f0c98405c65b8bdb597807c5fa84))
+
+
+### Features
+
+* [TECH-620]: sort rules engine events by date ([#1970](https://github.com/dhis2/capture-app/issues/1970)) ([b21d189](https://github.com/dhis2/capture-app/commit/b21d189b0e2372a34a44e084ead0ae157d42608f))
+
+## [1.37.2](https://github.com/dhis2/capture-app/compare/v1.37.1...v1.37.2) (2021-08-12)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([b9da318](https://github.com/dhis2/capture-app/commit/b9da31866b9fd845082a6c731827adb1aedddce0))
+
+## [1.37.1](https://github.com/dhis2/capture-app/compare/v1.37.0...v1.37.1) (2021-08-11)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([81898f1](https://github.com/dhis2/capture-app/commit/81898f1655b9a5de7d340032e28cfe86f5e85bd8))
+
+# [1.37.0](https://github.com/dhis2/capture-app/compare/v1.36.0...v1.37.0) (2021-08-10)
+
+
+### Features
+
+* [DHIS2-10222] Stages & Events ([#1788](https://github.com/dhis2/capture-app/issues/1788)) ([5240082](https://github.com/dhis2/capture-app/commit/52400826d2077ab113cbf99f299e2aa7106d9610)), closes [#1870](https://github.com/dhis2/capture-app/issues/1870) [#1875](https://github.com/dhis2/capture-app/issues/1875)
+
+# [1.36.0](https://github.com/dhis2/capture-app/compare/v1.35.1...v1.36.0) (2021-08-10)
+
+
+### Features
+
+*  [DHIS2-11573] Sort indicators alphabetically in indicator widget ([cc79744](https://github.com/dhis2/capture-app/commit/cc79744f07e449af862dc20af811dc1d629ce26e))
+
+## [1.35.1](https://github.com/dhis2/capture-app/compare/v1.35.0...v1.35.1) (2021-08-10)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([d3b53d7](https://github.com/dhis2/capture-app/commit/d3b53d7e67d32a1e3db60b9e9b056d681f8f3a5e))
+
+# [1.35.0](https://github.com/dhis2/capture-app/compare/v1.34.5...v1.35.0) (2021-08-06)
+
+
+### Features
+
+* [DHIS2-11208] Enrolment event form using the view/edit single event component ([#1803](https://github.com/dhis2/capture-app/issues/1803)) ([2bc5d2d](https://github.com/dhis2/capture-app/commit/2bc5d2d8659c6e7249888cbed4852a1a67492b3c))
+
+## [1.34.5](https://github.com/dhis2/capture-app/compare/v1.34.4...v1.34.5) (2021-07-28)
+
+
+### Bug Fixes
+
+* **translations:** sync translations from transifex (master) ([eafe60f](https://github.com/dhis2/capture-app/commit/eafe60fb83f99dfcd6849b0bebdb39611048a2ab))
+
 ## [1.34.4](https://github.com/dhis2/capture-app/compare/v1.34.3...v1.34.4) (2021-07-19)